+++ core/components/romanescobackyard/model/romanescobackyard/romanesco.class.php
@@ -1,7 +1,9 @@
 <?php
 /**
+ * @property modX modx
  * @package romanesco
  */
+
 class Romanesco
 {
     public $config = array();
@@ -33,5 +35,37 @@ class Romanesco
         }
         return $result;
         //yield $result;
+    }
+
+    // Generate critical CSS for a given page
+    public function generateCriticalCSS($resource) {
+        $cssPathSystem = $this->modx->getObject('modSystemSetting', array(
+            'key' => 'romanesco.custom_css_path'
+        ));
+        $cssPathContext = $this->modx->getObject('modContextSetting', array(
+            'context_key' => $resource->get('context_key'),
+            'key' => 'romanesco.custom_css_path'
+        ));
+
+        if ($cssPathContext) {
+            $cssPath = $this->modx->getOption('base_path') . $cssPathContext->get('value');
+        } else if ($cssPathSystem) {
+            $cssPath = $this->modx->getOption('base_path') . $cssPathSystem->get('value');
+        } else {
+            $cssPath = $this->modx->getOption('base_path') . 'assets/css';
+        }
+
+        $buildCommand = 'gulp critical --src ' . $this->modx->makeUrl($resource->get('id'),'','','full') . ' --dist ' . $cssPath . '/critical/'. rtrim($resource->get('uri'),'/') . '.css';
+
+        exec(
+            '"$HOME/.nvm/nvm-exec" ' . $buildCommand .
+            ' --gulpfile ' . escapeshellcmd($this->modx->getOption('assets_path')) . 'components/romanescobackyard/js/gulp/generate-critical-css.js' .
+            ' >> ' . escapeshellcmd($this->modx->getOption('core_path')) . 'cache/logs/css-critical.log' .
+            ' 2>>' . escapeshellcmd($this->modx->getOption('core_path')) . 'cache/logs/css-error.log &',
+            $output,
+            $return_css
+        );
+
+        return true;
     }
 }